Comments, refactoring and reformatting
