--- src/main/java/gui/realms/RealmState.java
@@ -1,8 +1,0 @@
-package gui.realms;
-
-public enum RealmState {
-    PLACEHOLDER,
-    LOADING,
-    IP_UNKNOWN,
-    IP_KNOWN
-}
