+++ src/main/resources/ui/RealmsTab.fxml
@@ -5,7 +5,7 @@
 
 <AnchorPane xmlns="http://javafx.com/javafx"
             xmlns:fx="http://javafx.com/fxml"
-            fx:controller="gui.realms.RealmsTabController"
+            fx:controller="gui.RealmsTabController"
             prefHeight="340.0" prefWidth="600.0">
         <ListView fx:id="serverList" layoutX="14.0" layoutY="14.0" prefHeight="160.0" prefWidth="580.0"/>
 </AnchorPane>
